remove multicoin/deprecated apis
