+++ plugin/evm/atomic/export_tx.go
@@ -287,7 +287,6 @@ func NewExportTx(
 	ctx *snow.Context,
 	rules params.Rules,
 	state StateDB,
-	assetID ids.ID, // AssetID of the tokens to export
 	amount uint64, // Amount of tokens to export
 	chainID ids.ID, // Chain to send the UTXOs to
 	to ids.ShortID, // Address of chain recipient
@@ -295,7 +294,7 @@ func NewExportTx(
 	keys []*secp256k1.PrivateKey, // Pay the fee and provide the tokens
 ) (*Tx, error) {
 	outs := []*avax.TransferableOutput{{
-		Asset: avax.Asset{ID: assetID},
+		Asset: avax.Asset{ID: ctx.AVAXAssetID},
 		Out: &secp256k1fx.TransferOutput{
 			Amt: amount,
 			OutputOwners: secp256k1fx.OutputOwners{
@@ -307,21 +306,11 @@ func NewExportTx(
 	}}
 
 	var (
-		avaxNeeded           uint64 = 0
+		avaxNeeded           uint64 = amount
 		ins, avaxIns         []EVMInput
 		signers, avaxSigners [][]*secp256k1.PrivateKey
 		err                  error
 	)
-
-	// consume non-AVAX
-	if assetID != ctx.AVAXAssetID {
-		ins, signers, err = GetSpendableFunds(ctx, state, keys, assetID, amount)
-		if err != nil {
-			return nil, fmt.Errorf("couldn't generate tx inputs/signers: %w", err)
-		}
-	} else {
-		avaxNeeded = amount
-	}
 
 	switch {
 	case rules.IsApricotPhase3:
@@ -350,7 +339,7 @@ func NewExportTx(
 		if err != nil {
 			return nil, errOverflowExport
 		}
-		avaxIns, avaxSigners, err = GetSpendableFunds(ctx, state, keys, ctx.AVAXAssetID, newAvaxNeeded)
+		avaxIns, avaxSigners, err = GetSpendableFunds(ctx, state, keys, newAvaxNeeded)
 	}
 	if err != nil {
 		return nil, fmt.Errorf("couldn't generate tx inputs/signers: %w", err)
@@ -420,7 +409,6 @@ func GetSpendableFunds(
 	ctx *snow.Context,
 	state StateDB,
 	keys []*secp256k1.PrivateKey,
-	assetID ids.ID,
 	amount uint64,
 ) ([]EVMInput, [][]*secp256k1.PrivateKey, error) {
 	inputs := []EVMInput{}
@@ -433,13 +421,10 @@ func GetSpendableFunds(
 		}
 		addr := GetEthAddress(key)
 		var balance uint64
-		if assetID == ctx.AVAXAssetID {
-			// If the asset is AVAX, we divide by the x2cRate to convert back to the correct
-			// denomination of AVAX that can be exported.
-			balance = new(uint256.Int).Div(state.GetBalance(addr), X2CRate).Uint64()
-		} else {
-			balance = state.GetBalanceMultiCoin(addr, common.Hash(assetID)).Uint64()
-		}
+		// we divide by the x2cRate to convert back to the correct
+		// denomination of AVAX that can be exported.
+		balance = new(uint256.Int).Div(state.GetBalance(addr), X2CRate).Uint64()
+
 		if balance == 0 {
 			continue
 		}
@@ -451,7 +436,7 @@ func GetSpendableFunds(
 		inputs = append(inputs, EVMInput{
 			Address: addr,
 			Amount:  balance,
-			AssetID: assetID,
+			AssetID: ctx.AVAXAssetID,
 			Nonce:   nonce,
 		})
 		signers = append(signers, []*secp256k1.PrivateKey{key})
